requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
Flask==2.3.2
frozenlist==1.4.0
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.0.2
MarkupSafe==2.1.3
msgpack==1.0.5
multidict==6.0.4
openai==0.27.8
packaging==23.1
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.65.0
urllib3==2.0.4
Werkzeug==2.3.6
yarl==1.9.2
